Move macro properties calculator to separate class
