+++ api/c/indigo/src/indigo_misc.cpp
@@ -23,6 +23,7 @@
 #include "molecule/icm_loader.h"
 #include "molecule/icm_saver.h"
 #include "molecule/ket_document_json_saver.h"
+#include "molecule/macro_properties_calculator.h"
 #include "molecule/molecule_arom.h"
 #include "molecule/molecule_automorphism_search.h"
 #include "molecule/molecule_hash.h"
@@ -1711,7 +1712,8 @@ CEXPORT const char* indigoMacroProperties(int object)
         if (IndigoBaseMolecule::is(obj) || IndigoBaseReaction::is(obj) || IndigoKetDocument::is(obj))
         {
             auto& doc = obj.getKetDocument();
-            doc.CalculateMacroProps(out, self.json_saving_pretty);
+            MacroPropertiesCalculator calc;
+            calc.CalculateMacroProps(doc, out, self.json_saving_pretty);
         }
         out.writeChar(0);
         return tmp.string.ptr();
